+++ src/interfaces/pool/IPoolStates.sol
@@ -7,7 +7,16 @@ interface IPoolStates {
     function poolInfo(address)
         external
         view
-        returns (uint256 reserveD, uint256 poolOwnershipUnitsTotal, uint256 reserveA, uint256 minLaunchReserveA, uint256 minLaunchReserveD, uint256 initialDToMint, uint256 poolFeeCollected, bool initialized);
+        returns (
+            uint256 reserveD,
+            uint256 poolOwnershipUnitsTotal,
+            uint256 reserveA,
+            uint256 minLaunchReserveA,
+            uint256 minLaunchReserveD,
+            uint256 initialDToMint,
+            uint256 poolFeeCollected,
+            bool initialized
+        );
     function userLpUnitInfo(address, address) external view returns (uint256);
     function VAULT_ADDRESS() external view returns (address);
     function ROUTER_ADDRESS() external view returns (address);
@@ -15,7 +24,7 @@ interface IPoolStates {
     function pairSlippage(bytes32) external view returns (uint256);
     function globalSlippage() external view returns (uint256);
     // function pairSwapHistory(bytes32) external view returns(uint256,uint256);
-    function poolStreamQueue(bytes32) external view returns(Swap[] memory swaps,uint front, uint back);
-    function pairStreamQueue(bytes32) external view returns(Swap[] memory swaps,uint front, uint back);
-    function pairPendingQueue(bytes32) external view returns(Swap[] memory swaps, uint front, uint back);
+    function poolStreamQueue(bytes32) external view returns (Swap[] memory swaps, uint256 front, uint256 back);
+    function pairStreamQueue(bytes32) external view returns (Swap[] memory swaps, uint256 front, uint256 back);
+    function pairPendingQueue(bytes32) external view returns (Swap[] memory swaps, uint256 front, uint256 back);
 }
